# Conflicts:
#	ThirdPlace-Backend/src/main/java/com/CherrySystems/ThirdPlace_Backend/models/User.java

+++ ThirdPlace-Backend/src/main/java/com/CherrySystems/ThirdPlace_Backend/controllers/SubmissionController.java
@@ -1,8 +1,11 @@
 package com.CherrySystems.ThirdPlace_Backend.controllers;
 
+import com.CherrySystems.ThirdPlace_Backend.models.Category;
+import com.CherrySystems.ThirdPlace_Backend.models.Review;
 import com.CherrySystems.ThirdPlace_Backend.models.Submission;
 import com.CherrySystems.ThirdPlace_Backend.models.User;
 import com.CherrySystems.ThirdPlace_Backend.models.dto.SubmissionFormDTO;
+import com.CherrySystems.ThirdPlace_Backend.repositories.CategoryRepository;
 import com.CherrySystems.ThirdPlace_Backend.repositories.SubmissionRepository;
 import com.CherrySystems.ThirdPlace_Backend.repositories.UserRepository;
 import jakarta.persistence.Id;
@@ -29,6 +32,9 @@ public class SubmissionController {
     private UserRepository userRepository;
 
     @Autowired
+    private CategoryRepository categoryRepository;
+
+    @Autowired
     private AuthenticationController authenticationController;
 
 
@@ -44,6 +50,9 @@ public class SubmissionController {
         newSubmission.setRating(submissionFormDTO.getRating());
         newSubmission.setDescription(submissionFormDTO.getDescription());
         newSubmission.setSubmissionReview(submissionFormDTO.getSubmissionReview());
+
+        List<Category> categoryList = (List<Category>) categoryRepository.findAllById(submissionFormDTO.getCategories());
+        newSubmission.setCategories(categoryList);
 
         //TODO: get user from session/authentication, currently holding dummy data to appease constructor
         User user = authenticationController.getUserFromSession(session);
@@ -91,7 +100,7 @@ public class SubmissionController {
 
     //Allows users to edit submission entity in DB
     @PatchMapping("/{id}")
-    public ResponseEntity<?> updateSubmission(@PathVariable Integer id, @RequestBody SubmissionFormDTO submissionFormDTO) {
+    public ResponseEntity<?> updateSubmission(@PathVariable Integer id, @RequestBody SubmissionFormDTO submissionFormDTO, HttpSession session) {
 
         //Finds submission by ID in repository
         Submission findInRepo = submissionRepository.findById(id).get();
@@ -103,9 +112,13 @@ public class SubmissionController {
         findInRepo.setDescription(submissionFormDTO.getDescription());
         findInRepo.setSubmissionReview(submissionFormDTO.getSubmissionReview());
 
+        List<Category> categoryList = (List<Category>) categoryRepository.findAllById(submissionFormDTO.getCategories());
+        findInRepo.setCategories(categoryList);
+
         //TODO: get user from session/authentication, currently holding dummy data to appease constructor
-        User user1 = userRepository.findByUsername("user1");
-        findInRepo.setUser(user1);
+        User user = authenticationController.getUserFromSession(session);
+//        User user1 = userRepository.findByUsername("user1");
+        findInRepo.setUser(user);
 
         //TODO: get placeID from Maps API address, currently holding dummy data to appease constructor
         findInRepo.setPlaceId("123abc");
@@ -118,7 +131,22 @@ public class SubmissionController {
 
     //Deletes submissions in Submission Repository by finding the submission by its ID#
     @DeleteMapping("/{id}")
-    public void deleteSubmission(@PathVariable Integer id) {
-        submissionRepository.deleteById(id);
+    public ResponseEntity<?> deleteSubmission(@PathVariable Integer id, HttpSession session) {
+
+        User user = authenticationController.getUserFromSession(session);
+        if (user == null) {
+            return ResponseEntity.badRequest().body("User is not logged in.");
+        }
+
+        //Find submission by ID in repository
+        Submission submissionById = submissionRepository.findById(id).get();
+
+        if (!(submissionById.getUser()).equals(user)) {
+            return ResponseEntity.badRequest().body("Submission can only be deleted by posting user.");
+        } else {
+            submissionRepository.deleteById(id);
+        }
+        return  ResponseEntity.ok("Submission deleted.");
     }
+
 }
